(doc) add required first argument to "init" method
As defined in the "www/PedometerCordova.js" file, the init method expect 2 arguments. The first is a string, the second is a callback. I have updated the code example in the README.md to include an empty string as the first argument. Without that, the callback can't ever be executed.